requirements.txt
# Python 3.10+
Django>=5
djangorestframework>=3.20
torch>=2.1
transformers>=4.50
bitsandbytes>=0.41
ekphrasis>=0.6
clean-text[emoji]>=0.6
